Work on sample rate independency

+++ hardware/AFSK.h
@@ -120,7 +120,7 @@ typedef struct Afsk
     int16_t iirX[2];                        // IIR Filter X cells
     int16_t iirY[2];                        // IIR Filter Y cells
 
-    uint8_t sampledBits;                    // Bits sampled by the demodulator (at ADC speed)
+    uint32_t sampledBits;                    // Bits sampled by the demodulator (at ADC speed)
     int8_t currentPhase;                    // Current phase of the demodulator
     uint8_t actualBits;                     // Actual found bits at correct bitrate
